+++ packages/shared/src/api/api.ts
@@ -24,17 +24,17 @@ const createApi = (baseUrl: string): JsonApiClient => {
 export const coreApi = createApi(CONFIG.CORE_API_BASE_URL)
 export const dexApi = createApi(CONFIG.DEX_API_BASE_URL)
 
-export const loadDataViaLoop = async <T>(
+export const loadAllPagesData = async <T extends Array<unknown>>(
   response: JsonApiResponse<T>,
   defaultPageLimit = CONFIG.MAX_PAGE_LIMIT,
 ): Promise<T> => {
-  let data = response.data as unknown as Array<T>
+  const data = response.data
   const pageLimit = response.pageLimit || defaultPageLimit
 
   while ((response.data as unknown as Array<T>).length === pageLimit) {
     response = await response.fetchPage(JsonApiLinkFields.next)
     const newResponseData = response.data as unknown as Array<T>
-    data = [...data, ...newResponseData]
+    data.push(...newResponseData)
   }
 
   return data as unknown as T