+++ packages/shared/src/api/id.ts
@@ -1,9 +1,5 @@
 import type { InternalTokenId } from '@/types'
 
-const parseID = <T extends [string | number, string]>(id: string): T => {
-  return id.split(':') as T
-}
-
 export const parseTokenId = (id: string): InternalTokenId => {
-  return parseID<InternalTokenId>(id)
+  return id.split(':') as InternalTokenId
 }